+++ src/extension/fixes/wrongIntellisenseMode.ts
@@ -1,6 +1,6 @@
 // M1 Macs have the wrong Intellisense mode set which causes errors
 
-import * as os from "os";
+import * as vscode from "vscode";
 
 import { ProjectUE4 } from "../../project/projectUE4";
 import * as consts from "../../consts";
@@ -8,15 +8,36 @@ import { isMacM1, setIntellisenseMode } from "../../shared";
 
 import * as console from "../../console";
 
+const INTELLISENSE_MODE = "intellisenseMode"
 
 export function fixWrongIntellisenseMode(project: ProjectUE4) {
     console.log("Start fix wrong intellisense mode");
+
+    const compilerConfig = vscode.workspace.getConfiguration(consts.CONFIG_SECTION_EXTENSION_COMPILER)
+    if(!compilerConfig){
+        console.error("Couldn't get Compiler Config!")
+        return;
+    }
+    let intellisenseMode = compilerConfig.get<string>(INTELLISENSE_MODE);
     
-    if(!isMacM1()){
-        console.log("Intellisense mode doesn't need to be fixed.");
+    if(isMacM1() && !intellisenseMode){
+        try {
+            compilerConfig.update(INTELLISENSE_MODE, consts.INTELLISENSE_MODE_CLANG_X64)
+        } catch (error) {
+            console.error("Couldn't update extension setting intellisenseMode!")
+            return;
+        }
+        
+        intellisenseMode = consts.INTELLISENSE_MODE_CLANG_X64;
         return;
     }
 
-    setIntellisenseMode(project, consts.INTELLISENSE_MODE_MACOS_CLANG_ARM64);
+    if(intellisenseMode){
+        setIntellisenseMode(project, intellisenseMode)
+        return;
+    }
+    else{
+        console.log("Didn't need to set intellisenseMode.")
+    }
 
 }